Zoomable spectrograms Decrease the IndexCalculation Duration down to 0.25 seconds. Works OK. Start work on compression of normal spectrograms. Original hash: 0d0daf63f9ec134f3f9b4f70ead54265b8bffda0

+++ AudioAnalysis/AnalysisPrograms/AnalyseLongRecordings/AnalyseLongRecording.cs
@@ -162,7 +162,7 @@ Output  to  directory: {1}
             // set IndexCalculationDuration i.e. duration of a subsegment
             try
             {
-                int indexCalculationDuration = configuration[AnalysisKeys.IndexCalculationDuration];
+                double indexCalculationDuration = configuration[AnalysisKeys.IndexCalculationDuration];
                 analysisSettings.IndexCalculationDuration = TimeSpan.FromSeconds(indexCalculationDuration);
             }
             catch (Exception ex)
